+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/service/ProtectorSupportMaintainServiceImpl.java
@@ -251,7 +251,7 @@ public class ProtectorSupportMaintainServiceImpl implements ProtectorSupportMain
             )).collect(Collectors.toList());
             crudService.batchInsert(protectorSupports);
         } catch (Exception e) {
-            throw ServiceExceptionHelper.logAndThrow("重置调度器支持时发生异常",
+            throw ServiceExceptionHelper.logAndThrow("重置保护器支持时发生异常",
                     LogLevel.WARN, sem, e
             );
         }