+++ account-keeper-impl/src/test/java/com/dwarfeng/acckeeper/impl/configuration/ServiceExceptionMapperConfiguration.java
@@ -23,6 +23,11 @@ public class ServiceExceptionMapperConfiguration {
         destination.put(LoginStateNotExistsException.class, ServiceExceptionCodes.LOGIN_STATE_NOT_EXISTS);
         destination.put(PasswordIncorrectException.class, ServiceExceptionCodes.PASSWORD_INCORRECT);
         destination.put(SerialVersionInconsistentException.class, ServiceExceptionCodes.SERIAL_VERSION_INCONSISTENT);
+        destination.put(ProtectorException.class, ServiceExceptionCodes.PROTECTOR_FAILED);
+        destination.put(ProtectorExecutionException.class, ServiceExceptionCodes.PROTECTOR_EXECUTION_FAILED);
+        destination.put(ProtectorMakeException.class, ServiceExceptionCodes.PROTECTOR_MAKE_FAILED);
+        destination.put(UnsupportedProtectorTypeException.class, ServiceExceptionCodes.UNSUPPORTED_PROTECTOR_TYPE);
+        destination.put(ProtectorInfoNotExistsException.class, ServiceExceptionCodes.PROTECTOR_INFO_NOT_EXISTED);
         return new MapServiceExceptionMapper(destination, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINE);
     }
 }