+++ account-keeper-sdk/src/main/java/com/dwarfeng/acckeeper/sdk/util/ServiceExceptionCodes.java
@@ -12,41 +12,30 @@ public final class ServiceExceptionCodes {
 
     private static int EXCEPTION_CODE_OFFSET = 1000;
 
-    /**
-     * 账户已经存在。
-     */
     public static final ServiceException.Code ACCOUNT_ALREADY_EXISTED =
             new ServiceException.Code(offset(0), "account already existed");
-    /**
-     * 账户禁用。
-     */
     public static final ServiceException.Code ACCOUNT_DISABLED =
             new ServiceException.Code(offset(10), "account disabled");
-    /**
-     * 账户不存在。
-     */
     public static final ServiceException.Code ACCOUNT_NOT_EXISTS =
             new ServiceException.Code(offset(20), "account not exists");
-    /**
-     * 登录超时。
-     */
     public static final ServiceException.Code LOGIN_STATE_EXPIRED =
             new ServiceException.Code(offset(30), "login state expired");
-    /**
-     * 登录超时。
-     */
     public static final ServiceException.Code LOGIN_STATE_NOT_EXISTS =
             new ServiceException.Code(offset(40), "login state not exists");
-    /**
-     * 密码错误。
-     */
     public static final ServiceException.Code PASSWORD_INCORRECT =
             new ServiceException.Code(offset(50), "password incorrect");
-    /**
-     * 登录过期。
-     */
     public static final ServiceException.Code SERIAL_VERSION_INCONSISTENT =
             new ServiceException.Code(offset(60), "serial version inconsistent");
+    public static final ServiceException.Code PROTECTOR_FAILED =
+            new ServiceException.Code(offset(70), "protector failed");
+    public static final ServiceException.Code PROTECTOR_EXECUTION_FAILED =
+            new ServiceException.Code(offset(71), "protector execution failed");
+    public static final ServiceException.Code PROTECTOR_MAKE_FAILED =
+            new ServiceException.Code(offset(72), "protector make failed");
+    public static final ServiceException.Code UNSUPPORTED_PROTECTOR_TYPE =
+            new ServiceException.Code(offset(73), "unsupported protector type");
+    public static final ServiceException.Code PROTECTOR_INFO_NOT_EXISTED =
+            new ServiceException.Code(offset(80), "protector info not existed");
 
     private static int offset(int i) {
         return EXCEPTION_CODE_OFFSET + i;
@@ -67,7 +56,22 @@ public final class ServiceExceptionCodes {
      * @param exceptionCodeOffset 指定的异常代号的偏移量。
      */
     public static void setExceptionCodeOffset(int exceptionCodeOffset) {
+        // 设置 EXCEPTION_CODE_OFFSET 的值。
         EXCEPTION_CODE_OFFSET = exceptionCodeOffset;
+
+        // 以新的 EXCEPTION_CODE_OFFSET 为基准，更新异常代码的值。
+        ACCOUNT_ALREADY_EXISTED.setCode(offset(0));
+        ACCOUNT_DISABLED.setCode(offset(10));
+        ACCOUNT_NOT_EXISTS.setCode(offset(20));
+        LOGIN_STATE_EXPIRED.setCode(offset(30));
+        LOGIN_STATE_NOT_EXISTS.setCode(offset(40));
+        PASSWORD_INCORRECT.setCode(offset(50));
+        SERIAL_VERSION_INCONSISTENT.setCode(offset(60));
+        PROTECTOR_FAILED.setCode(offset(70));
+        PROTECTOR_EXECUTION_FAILED.setCode(offset(71));
+        PROTECTOR_MAKE_FAILED.setCode(offset(72));
+        UNSUPPORTED_PROTECTOR_TYPE.setCode(offset(73));
+        PROTECTOR_INFO_NOT_EXISTED.setCode(offset(80));
     }
 
     private ServiceExceptionCodes() {